requirements.txt
Django==2.0.8
django-bootstrap-breadcrumbs==0.9.1
django-likert-field==0.2.0
django-recaptcha==1.4.0
easy-thumbnails==2.5
Pillow==5.2.0
Unidecode==1.0.22
